Release 0.13: fixed concat_pat #26
concat patterns together, and stop using our engine
with the first pattern which is not ours or core.
re_compile to core and fallback to core then.

Problem is when core fails to concat a plugin pattern.
(even if it should not. the PV and flags should be enough, but this
is a core bug we have to work against).