Avoid a race in t/gh29-segv.t
If the a parent process was too slow, a child process could print
"ok 2" before the parent printed "ok 1". That resulted into unordered
tests and was reported as an error.

This patch eliminites the race by synchronizing on the child
termination by calling waitpid(). Nevertheless it's necessary to
collect child's statuses to prevent from creating zombies.

https://github.com/rurban/re-engine-PCRE2/issues/34