PCRE2.xs - set up logical_parens on 5.37.8, and related fixes
logical_parens needs to be set up the same as nparens unless
branch reset is in use (?|(...)|(...)) buffers are now separated
into logical buffers (eg $1) and physical buffers. In the case
of a branch reset like above there would be two physical buffers,
and 1 logical buffer.

I dont think PCRE2 does branch reset properly, so we can just
default it to the same value of nparens.

This patch also includes changes to set re->lastparen and
re->lastcloseparen in the exec sub, and not in the compile sub.
Last paren indicates the highest capture buffer that has content,
and lastcloseparen indicates the last capture buffer that match
so that we can support $^N. These two values are subtly different
and are not necessarily the same, although usually they are.