PCRE2.xs - use savepvn to copy the matched string
This is probably not the most efficient way to deal with this,
but we cant just pointer copy the buffer. It might be modified
or freed while the regex is still alive, and then bad things happen
when the regex is freed. Instead we just clone it with savepvn()

Ideally we would more closely mimic the behavior of core in this
regard, so that we can take advantage of perl level COW support
to get a cheaper string clone.